finite-field: implement finite field arithmetic
Implement the finite field arithmetic needed for BLS12-381. This
includes arithmetic in the field F_p, F_p^2, F_p^6, and F_p^12, where
the field extensions are implemented as tower fields using the
algorithms described in "Guide to Pairing-based Cryptography."

Add some initial tests to validate these implementations. At this point,
all tests are passing, and the finite field arithmetic should provide a
good enough base for the next step.

Some improvements will need to be made, such as removing the need to
pass p around, and instead declare it as a static global.

Signed-off-by: Nick Rosbrook <nr@enr0n.net>