finite-field: add static global g_BLS12_381_P
Add a static global to finite_field.c to hold the field characteristic P
so that it does not need to be passed around every function. At the
moment, since all of the finite field arithmetic contains BLS12-381
specific code anyways, there is no harm in always setting this value to
BLS12-381's prime P.

The variable must be initialized with fp_params_init(), and free'd with
fp_params_free().

Signed-off-by: Nick Rosbrook <nr@enr0n.net>