G2: implement basic needs for G_2 group
This is very much a clone of 9fd3cf2. The same algorithms are used, and
an equivalent API is exposed in G2.h as G1.h. The only real differences
between the two are that G_2 is of coursed defined as a subgroup of the
twist E': y^2 = x^3 + 4(u+1) defined over F_p^2. So, the G2.c
implementation works with the fp2 API rather than libgmp directly.

The tests are basically the same too, with updated test vectors.

Signed-off-by: Nick Rosbrook <nr@enr0n.net>