pairing: implement final exponentiation
Somehow, it made more sense to implement and test this first. Use the
theorem from the "homogeneous cyclotomic polynomial" paper to implement
the hard part of the final exponentiation. This provides a pretty
straight-forward approach to final exponentiation.

To support the algorithm, the frobenious endomorphism is needed, in
addition to plain exponentiation in Fp^12. For the former, we used
pre-computed coeffients that are constant with respect to the field
characteristic p. For the latter, we use a simple square-and-multiply
algorithm.

Each of these additions are verified with tests.

Signed-off-by: Nick Rosbrook <nr@enr0n.net>