pairing: implement line functions to complete ate pairing
Complete the line functions using to complete the ate pairing
implementation. Note that we use the method of computing the pairing on
the twist, i.e. twist the point from G_1 rather than untwisting the
point from G_2.

With this, the ate test case is passing. Some tests should still be
added to check bilinearity.

Signed-off-by: Nick Rosbrook <nr@enr0n.net>