pairing: add miller loop point accumulation
Add the basic skeleton of the miller loop, includig the point doubling
and addition. This allows a unit test to verify the iteration over the
curve parameter t in binary, and also that the correct points in G2 are
being used to calculate the line functions.

N.B. that the expected output for the miller loop test may not be
correct. It is expected to be since the final exp of that output leads
to the expected pairing of our generators. But, this should be double
checked when line functions are implemented.

Signed-off-by: Nick Rosbrook <nr@enr0n.net>