hash: implement hash_to_field_fp
Implement the hash_to_field function for fp, which is needed for G1
hashing. This requires adding a substr function for octet_string, and
fixing the lack of const in octet_string_to_str, and adding OS2IP.

The OS2IP implementation is sort of lazy, but it works. It should be
re-visited at a later date to get rid of the string usage.

Finally, add tests for the vectors from [1].

[1] https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-11#appendix-J.9.1

Signed-off-by: Nick Rosbrook <nr@enr0n.net>