G1/G2: fix scalar multiplication with negatives
The G1 and G2 scalar multiplication functions do not currently check if
the scalar m is negative. Add logic to fix this, and while here, add
functions to use projective coordinates directly for scalar
multiplication.

At a later date, the API naming for these functions should be cleaned
up, and _affine versions of add functions should probably be added.

Signed-off-by: Nick Rosbrook <nr@enr0n.net>